Move Debug code paths from the main scene delegate to a new debug scene class
